requirements.txt
fastapi
uvicorn
sqlmodel
pydantic
paramiko
passlib
jwt
python-multipart
python-jose[cryptography]
passlib[bcrypt]